backend/requirements.txt
# Web 框架
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9  # 文件上传

# 配置管理
pydantic-settings>=2.2.0

# 数据库
sqlalchemy>=2.0.30
alembic>=1.13.0

# 任务调度（APScheduler 替代 Celery/Redis，无需外部服务）
APScheduler>=3.10.0

# Instagram 自动化
playwright>=1.44.0

# YouTube API
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# HTTP 请求
requests>=2.31.0

# JWT
PyJWT>=2.8.0

# 工具
python-dotenv>=1.0.0
cryptography>=42.0.0
instagrapi>=2.1.2
